+++ topobank_publication/signals.py
@@ -3,13 +3,20 @@ from django.dispatch import receiver
 import short_url
 import logging
 
-from .models import Publication
+from .models import Publication, PublicationCollection
 
 _log = logging.getLogger(__name__)
 
 
 @receiver(post_save, sender=Publication)
 def set_short_url(sender, instance, **kwargs):
+    if instance.short_url is None:
+        instance.short_url = short_url.encode_url(instance.id)
+        instance.save()
+
+
+@receiver(post_save, sender=PublicationCollection)
+def set_short_url_publication_collection(sender, instance, **kwargs):
     if instance.short_url is None:
         instance.short_url = short_url.encode_url(instance.id)
         instance.save()